refactor(将对象相关函数移动至单独的文件):
