+++ lib/config.js
@@ -2,8 +2,8 @@
 
 import BLOG from '@/blog.config'
 import { useGlobal } from './global'
-import { deepClone } from './utils'
 import { isUrl } from '@/lib/utils/url'
+import { deepClone } from '@/lib/utils/objectUtils'
 
 /**
  * 读取配置顺序
