+++ lib/db/getSiteData.js
@@ -9,10 +9,10 @@ import getPageProperties, {
 } from '@/lib/notion/getPageProperties'
 import { fetchInBatches, getPage } from '@/lib/notion/getPostBlocks'
 import { compressImage, mapImgUrl } from '@/lib/notion/mapImage'
-import { deepClone } from '@/lib/utils'
 import { idToUuid } from 'notion-utils'
 import { siteConfig } from '../config'
 import { extractLangId, extractLangPrefix, getShortId } from '../utils/pageId'
+import { deepClone } from '@/lib/utils/objectUtils'
 
 export { getAllTags } from '../notion/getAllTags'
 export { getPost } from '../notion/getNotionPost'
