+++ lib/lang.js
@@ -1,5 +1,4 @@
 import BLOG from '@/blog.config'
-import { mergeDeep } from '@/lib/utils'
 import enUS from './lang/en-US'
 import frFR from './lang/fr-FR'
 import jaJP from './lang/ja-JP'
@@ -11,6 +10,7 @@ import { extractLangPrefix } from './utils/pageId'
 
 import { getQueryVariable } from '@/lib/utils/url'
 import { isBrowser } from '@/lib/utils/envUtils'
+import { mergeDeep } from '@/lib/utils/objectUtils'
 
 /**
  * 在这里配置所有支持的语言
