+++ lib/notion/getAllCategories.js
@@ -1,4 +1,4 @@
-import { isIterable } from '../utils'
+import { isIterable } from '@/lib/utils/objectUtils'
 
 /**
  * 获取所有文章的标签
