+++ lib/notion/getAllTags.js
@@ -1,5 +1,6 @@
 import { siteConfig } from '../config'
-import { isIterable } from '../utils'
+
+import { isIterable } from '@/lib/utils/objectUtils'
 
 /**
  * 获取所有文章的标签
