+++ lib/notion/getNotionConfig.js
@@ -7,9 +7,9 @@
  *
  */
 import { getDateValue, getTextContent } from 'notion-utils'
-import { deepClone } from '../utils'
 import getAllPageIds from './getAllPageIds'
 import { getPage } from './getPostBlocks'
+import { deepClone } from '@/lib/utils/objectUtils'
 
 /**
  * 从Notion中读取Config配置表
