+++ lib/notion/getPostBlocks.js
@@ -4,8 +4,9 @@ import {
   getOrSetDataWithCache,
   setDataToCache
 } from '@/lib/cache/cache_manager'
-import { deepClone, delay } from '../utils'
+import { delay } from '../utils'
 import notionAPI from '@/lib/notion/getNotionAPI'
+import { deepClone } from '@/lib/utils/objectUtils'
 
 /**
  * 获取文章内容块
