+++ lib/utils/objectUtils.js
@@ -1,0 +1,82 @@
+/**
+ * 打乱数组
+ * @param {*} array
+ * @returns
+ */
+export const shuffleArray = array => {
+  if (!array) {
+    return []
+  }
+  for (let i = array.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1))
+    ;[array[i], array[j]] = [array[j], array[i]]
+  }
+  return array
+}
+
+/**
+ * 深度合并两个对象
+ * @param target
+ * @param sources
+ */
+export function mergeDeep(target, ...sources) {
+  if (!sources.length) return target
+  const source = sources.shift()
+
+  if (isObject(target) && isObject(source)) {
+    for (const key in source) {
+      if (isObject(source[key])) {
+        if (!target[key]) Object.assign(target, { [key]: {} })
+        mergeDeep(target[key], source[key])
+      } else {
+        Object.assign(target, { [key]: source[key] })
+      }
+    }
+  }
+  return mergeDeep(target, ...sources)
+}
+
+/**
+ * 是否对象
+ * @param item
+ * @returns {boolean}
+ */
+export function isObject(item) {
+  return item && typeof item === 'object' && !Array.isArray(item)
+}
+
+/**
+ * 是否可迭代
+ * @param {*} obj
+ * @returns
+ */
+export function isIterable(obj) {
+  return obj != null && typeof obj[Symbol.iterator] === 'function'
+}
+
+/**
+ * 深拷贝对象
+ * 根据源对象类型深度复制，支持object和array
+ * @param {*} obj
+ * @returns
+ */
+export function deepClone(obj) {
+  if (Array.isArray(obj)) {
+    // If obj is an array, create a new array and deep clone each element
+    return obj.map(item => deepClone(item))
+  } else if (obj && typeof obj === 'object') {
+    const newObj = {}
+    for (const key in obj) {
+      if (Object.prototype.hasOwnProperty.call(obj, key)) {
+        if (obj[key] instanceof Date) {
+          newObj[key] = new Date(obj[key].getTime()).toISOString()
+        } else {
+          newObj[key] = deepClone(obj[key])
+        }
+      }
+    }
+    return newObj
+  } else {
+    return obj
+  }
+}
