+++ themes/game/components/GameListRealate.js
@@ -1,7 +1,7 @@
 /* eslint-disable @next/next/no-img-element */
-import { deepClone } from '@/lib/utils'
 import Link from 'next/link'
 import { useState } from 'react'
+import { deepClone } from '@/lib/utils/objectUtils'
 
 /**
  * 游戏列表- 关联游戏，在详情页展示
